backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
groq==0.4.1
langchain==0.3.18
langchain-groq==0.2.4
langchain-core==0.3.34
Faker==22.0.0
